feat: shared wrap helpers — GPUI cleanup + SwiftUI SBS wrap rendering (#54)
